feat: adding team title on leaderboard screen
