+++ screens/team/teamComponents/index.js
@@ -5,3 +5,4 @@ export { default as JoinTeamForm } from './JoinTeamForm';
 export { default as CreateTeamForm } from './CreateTeamForm';
 export { default as MemberCard } from './MemberCard';
 export { default as TeamListCard } from './TeamListCard';
+export { default as TeamTitle } from './TeamTitle';
